Need to fetch full repo with tags to get version for publishing
Versioneer relies on tags to create a version number. That means we need
the git repo to include tags. Github's checkout action changed to a
shallow checkout with only the latest commit, which does not have tags.
Easiest (maybe not only) fix is to do a 'git fetch --tags --unshallow'
to get a full copy of the repo with tags.